seperated buttons tests to see if one particular test fails to be run in travis.
